refactor(db): déplace les types tavern dans leur propre domaine
TavernConfig et TavernActivityType vivaient dans navigation/world/types.
Ils ont maintenant leur domaine packages/db/src/domains/tavern. navigation
(defineIsland, registry) importe le type depuis tavern ; TAVERN_BY_ZONE reste
construit dans le registry navigation car il dérive d'ISLAND_REGISTRY.
