+++ packages/db/src/domains/tavern/index.ts
@@ -1,0 +1,1 @@
+export { type TavernActivityType, type TavernConfig } from './types.js';
