+++ packages/db/src/index.ts
@@ -21,13 +21,12 @@ export {
   type Island,
   type Sea,
   type SubZone,
-  type TavernActivityType,
-  type TavernConfig,
   type TravelModifier,
   type Zone,
   TAVERN_BY_ZONE,
   WORLD_EDGES,
 } from './domains/navigation/index.js';
+export { type TavernActivityType, type TavernConfig } from './domains/tavern/index.js';
 export { ONBOARDING_STEP_IDS, type OnboardingStepId, type Player } from './domains/player/index.js';
 export { type DevilFruitTemplate } from './domains/devil_fruit/index.js';
 export {
